requirements.txt
langchain
deepseek
google-search-results
streamlit
python-dotenv
pandas
matplotlib
